share error type across mods
CMK-18553
